+++ src/interpreter/primitive-props.ts
@@ -120,6 +120,37 @@ const PRIMITIVE_PROPS: {
 			return Number.isNaN(res) ? NULL : NUM(res);
 		}),
 
+		starts_with: (target: VStr): VFn => FN_NATIVE(async ([prefix, start_index], _opts) => {
+			assertString(prefix);
+			if (!prefix.value) {
+				return TRUE;
+			}
+
+			if (start_index) assertNumber(start_index);
+			const raw_index = start_index?.value ?? 0;
+			if (raw_index < -target.value.length || raw_index > target.value.length) {
+				return FALSE;
+			}
+			const index = (raw_index >= 0) ? raw_index : target.value.length + raw_index;
+			return target.value.startsWith(prefix.value, index) ? TRUE : FALSE;
+		}),
+
+		ends_with: (target: VStr): VFn => FN_NATIVE(async ([suffix, end_index], _opts) => {
+			assertString(suffix);
+			if (!suffix.value) {
+				return TRUE;
+			}
+
+			if (end_index) assertNumber(end_index);
+			const raw_index = end_index?.value ?? target.value.length;
+			if (raw_index < -target.value.length || raw_index > target.value.length) {
+				return FALSE;
+			}
+			const index = (raw_index >= 0) ? raw_index : target.value.length + raw_index;
+
+			return target.value.endsWith(suffix.value, index) ? TRUE : FALSE;
+		}),
+
 		pad_start: (target: VStr): VFn => FN_NATIVE(([width, pad], _) => {
 			assertNumber(width);
 			const s = (pad) ? (assertString(pad), pad.value) : ' ';