+++ test/index.ts
@@ -2628,7 +2628,91 @@ describe('primitive props', () => {
 			);
 		});
 
-		test.concurrent("pad_start", async () => {
+		test.concurrent('starts_with (no index)', async () => {
+			const res = await exe(`
+			let str = "hello"
+			let empty = ""
+			<: [
+				str.starts_with(""), str.starts_with("hello"),
+				str.starts_with("he"), str.starts_with("ell"),
+				empty.starts_with(""), empty.starts_with("he"),
+			]
+			`);
+			eq(res, ARR([
+				TRUE, TRUE,
+				TRUE, FALSE,
+				TRUE, FALSE, 
+			]));
+		});
+
+		test.concurrent('starts_with (with index)', async () => {
+			const res = await exe(`
+			let str = "hello"
+			let empty = ""
+			<: [
+				str.starts_with("", 4), str.starts_with("he", 0),
+				str.starts_with("ll", 2), str.starts_with("lo", 3),
+				str.starts_with("lo", -2), str.starts_with("hel", -5),
+				str.starts_with("he", 2), str.starts_with("loa", 3),
+				str.starts_with("lo", -6), str.starts_with("", -7),
+				str.starts_with("lo", 6), str.starts_with("", 7),
+				empty.starts_with("", 2), empty.starts_with("ll", 2),
+			]
+			`);
+			eq(res, ARR([
+				TRUE, TRUE,
+				TRUE, TRUE,
+				TRUE, TRUE,
+				FALSE, FALSE,
+				FALSE, TRUE,
+				FALSE, TRUE,
+				TRUE, FALSE,
+			]));
+		});
+
+		test.concurrent('ends_with (no index)', async () => {
+			const res = await exe(`
+			let str = "hello"
+			let empty = ""
+			<: [
+				str.ends_with(""), str.ends_with("hello"),
+				str.ends_with("lo"), str.ends_with("ell"),
+				empty.ends_with(""), empty.ends_with("he"),
+			]
+			`);
+			eq(res, ARR([
+				TRUE, TRUE,
+				TRUE, FALSE,
+				TRUE, FALSE,
+			]));
+		});
+
+		test.concurrent('ends_with (with index)', async () => {
+			const res = await exe(`
+			let str = "hello"
+			let empty = ""
+			<: [
+				str.ends_with("", 3), str.ends_with("lo", 5),
+				str.ends_with("ll", 4), str.ends_with("he", 2),
+				str.ends_with("ll", -1), str.ends_with("he", -3),
+				str.ends_with("he", 5), str.ends_with("lo", 3),
+				str.ends_with("lo", -6), str.ends_with("", -7),
+				str.ends_with("lo", 6), str.ends_with("", 7),
+				empty.ends_with("", 2), empty.ends_with("ll", 2),
+			]
+			`);
+			eq(res, ARR([
+				TRUE, TRUE,
+				TRUE, TRUE,
+				TRUE, TRUE,
+				FALSE, FALSE,
+				FALSE, TRUE,
+				FALSE, TRUE,
+				TRUE, FALSE,
+			]));
+		});
+
+    test.concurrent("pad_start", async () => {
 			const res = await exe(`
 			let str = "abc"
 			<: [