+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - `arr.sort`の処理を非同期的にして高速化
 - `arr.flat`,`arr.flat_map`を追加
 - `Uri:encode_full`, `Uri:encode_component`, `Uri:decode_full`, `Uri:decode_component`を追加
+- `str.starts_with`,`str.ends_with`を追加
 
 # 0.18.0
 - `Core:abort`でプログラムを緊急停止できるように